+++ src/main/java/net/vandeneijk/shadowkickboxing/repositories/BodyPartRepository.java
@@ -7,6 +7,6 @@ package net.vandeneijk.shadowkickboxing.repositories;
 import net.vandeneijk.shadowkickboxing.models.BodyPart;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BodyPartRepository extends CrudRepository<BodyPart, Integer> {
+public interface BodyPartRepository extends CrudRepository<BodyPart, Long> {
 
 }
